OsLib_Geometry methods
