--- test/geometry/test_geometry_modify.rb
@@ -1,7 +1,0 @@
-require_relative '../helpers/minitest_helper'
-
-class TestGeometryModify < Minitest::Test
-  def setup
-    @geo = OpenstudioStandards::Geometry
-  end
-end
